requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
anthropic>=0.25.0
python-dotenv>=1.0.0
numpy>=1.24.0
fpdf2>=2.7.0
openpyxl>=3.1.0
